requirements.txt
torch==1.8.1
torchvision==0.9.1
numpy>=1.19.5
Pillow==8.0.1